requirements.txt
Django>=4.2.7,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.26.5,<0.27
gunicorn>=21.2.0,<21.3
django-cors-headers>=4.3.1,<4.4